requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
et-xmlfile==1.1.0
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.21.6
openai==0.25.0
openpyxl==3.0.10
pandas==1.3.5
pandas-stubs==1.2.0.62
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2020.1
requests==2.23.0
six==1.14.0
tqdm==4.64.1
twilio==6.40.0
typing-extensions==4.4.0
urllib3==1.25.11
Werkzeug==1.0.1
